requirements.txt
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
psycopg2==2.9.1
validate-email==1.3
whitenoise
xlwt==1.3.0
django-environ
weasyprint